% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBoot2.R
\name{myboot2}
\alias{myboot2}
\title{MyBoot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iteration}

\item{x}{x values}

\item{fun}{mean}

\item{...}{}
}
\value{

}
\description{
MyBoot2
}
