% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Central Limit Thoerem Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{interations}

\item{lambda}{lambda}

\item{...}{}
}
\value{
a barplot
}
\description{
My Central Limit Thoerem Poisson
}
