% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyMaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function type}

\item{x}{x as vector}

\item{param}{parameter}

\item{...}{}
}
\value{
list of values
}
\description{
Mymaxlik
}
